initial work on denyable

+++ lib/model/om/BaseContentObject.php
@@ -659,6 +659,11 @@ abstract class BaseContentObject extends BaseObject  implements Persistent
 			$deleteQuery = ContentObjectQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(ContentObjectPeer::isIgnoringRights() || ContentObjectPeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -709,6 +714,11 @@ abstract class BaseContentObject extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(ContentObjectPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(ContentObjectPeer::isIgnoringRights() || ContentObjectPeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -726,6 +736,11 @@ abstract class BaseContentObject extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(ContentObjectPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(ContentObjectPeer::isIgnoringRights() || ContentObjectPeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
