+++ lib/model/om/BaseDocument.php
@@ -1196,6 +1196,11 @@ abstract class BaseDocument extends BaseObject  implements Persistent
 			if(ReferencePeer::hasReference($this)) {
 				throw new PropelException("Exception in ".__METHOD__.": tried removing an instance from the database even though it is still referenced.");
 			}
+			// denyable behavior
+			if(!(DocumentPeer::isIgnoringRights() || DocumentPeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -1246,6 +1251,11 @@ abstract class BaseDocument extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(DocumentPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(DocumentPeer::isIgnoringRights() || DocumentPeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -1263,6 +1273,11 @@ abstract class BaseDocument extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(DocumentPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(DocumentPeer::isIgnoringRights() || DocumentPeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
