+++ lib/model/om/BaseDocumentCategory.php
@@ -682,6 +682,11 @@ abstract class BaseDocumentCategory extends BaseObject  implements Persistent
 			if(ReferencePeer::hasReference($this)) {
 				throw new PropelException("Exception in ".__METHOD__.": tried removing an instance from the database even though it is still referenced.");
 			}
+			// denyable behavior
+			if(!(DocumentCategoryPeer::isIgnoringRights() || DocumentCategoryPeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -732,6 +737,11 @@ abstract class BaseDocumentCategory extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(DocumentCategoryPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(DocumentCategoryPeer::isIgnoringRights() || DocumentCategoryPeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -749,6 +759,11 @@ abstract class BaseDocumentCategory extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(DocumentCategoryPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(DocumentCategoryPeer::isIgnoringRights() || DocumentCategoryPeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
