+++ lib/model/om/BaseGroupRole.php
@@ -501,6 +501,11 @@ abstract class BaseGroupRole extends BaseObject  implements Persistent
 			$deleteQuery = GroupRoleQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(GroupRolePeer::isIgnoringRights() || GroupRolePeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -551,6 +556,11 @@ abstract class BaseGroupRole extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(GroupRolePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(GroupRolePeer::isIgnoringRights() || GroupRolePeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -568,6 +578,11 @@ abstract class BaseGroupRole extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(GroupRolePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(GroupRolePeer::isIgnoringRights() || GroupRolePeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
