+++ lib/model/om/BaseLanguage.php
@@ -606,6 +606,11 @@ abstract class BaseLanguage extends BaseObject  implements Persistent
 			$deleteQuery = LanguageQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(LanguagePeer::isIgnoringRights() || LanguagePeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -656,6 +661,11 @@ abstract class BaseLanguage extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(LanguagePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(LanguagePeer::isIgnoringRights() || LanguagePeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -673,6 +683,11 @@ abstract class BaseLanguage extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(LanguagePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(LanguagePeer::isIgnoringRights() || LanguagePeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
