+++ lib/model/om/BaseLanguageObject.php
@@ -547,6 +547,11 @@ abstract class BaseLanguageObject extends BaseObject  implements Persistent
 			$deleteQuery = LanguageObjectQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(LanguageObjectPeer::isIgnoringRights() || LanguageObjectPeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -597,6 +602,11 @@ abstract class BaseLanguageObject extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(LanguageObjectPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(LanguageObjectPeer::isIgnoringRights() || LanguageObjectPeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -614,6 +624,11 @@ abstract class BaseLanguageObject extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(LanguageObjectPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(LanguageObjectPeer::isIgnoringRights() || LanguageObjectPeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
