+++ lib/model/om/BaseLanguageObjectHistory.php
@@ -584,6 +584,11 @@ abstract class BaseLanguageObjectHistory extends BaseObject  implements Persiste
 			$deleteQuery = LanguageObjectHistoryQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(LanguageObjectHistoryPeer::isIgnoringRights() || LanguageObjectHistoryPeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -634,6 +639,11 @@ abstract class BaseLanguageObjectHistory extends BaseObject  implements Persiste
 				if (!$this->isColumnModified(LanguageObjectHistoryPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(LanguageObjectHistoryPeer::isIgnoringRights() || LanguageObjectHistoryPeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -651,6 +661,11 @@ abstract class BaseLanguageObjectHistory extends BaseObject  implements Persiste
 				if ($this->isModified() && !$this->isColumnModified(LanguageObjectHistoryPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(LanguageObjectHistoryPeer::isIgnoringRights() || LanguageObjectHistoryPeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
