+++ lib/model/om/BaseLink.php
@@ -862,6 +862,11 @@ abstract class BaseLink extends BaseObject  implements Persistent
 			if(ReferencePeer::hasReference($this)) {
 				throw new PropelException("Exception in ".__METHOD__.": tried removing an instance from the database even though it is still referenced.");
 			}
+			// denyable behavior
+			if(!(LinkPeer::isIgnoringRights() || LinkPeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -912,6 +917,11 @@ abstract class BaseLink extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(LinkPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(LinkPeer::isIgnoringRights() || LinkPeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -929,6 +939,11 @@ abstract class BaseLink extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(LinkPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(LinkPeer::isIgnoringRights() || LinkPeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
