+++ lib/model/om/BasePageProperty.php
@@ -562,6 +562,11 @@ abstract class BasePageProperty extends BaseObject  implements Persistent
 			$deleteQuery = PagePropertyQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(PagePropertyPeer::isIgnoringRights() || PagePropertyPeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -612,6 +617,11 @@ abstract class BasePageProperty extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(PagePropertyPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(PagePropertyPeer::isIgnoringRights() || PagePropertyPeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -629,6 +639,11 @@ abstract class BasePageProperty extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(PagePropertyPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(PagePropertyPeer::isIgnoringRights() || PagePropertyPeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
