+++ lib/model/om/BasePageString.php
@@ -726,6 +726,11 @@ abstract class BasePageString extends BaseObject  implements Persistent
 			$deleteQuery = PageStringQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(PageStringPeer::isIgnoringRights() || PageStringPeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -776,6 +781,11 @@ abstract class BasePageString extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(PageStringPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(PageStringPeer::isIgnoringRights() || PageStringPeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -793,6 +803,11 @@ abstract class BasePageString extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(PageStringPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(PageStringPeer::isIgnoringRights() || PageStringPeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
