+++ lib/model/om/BaseReference.php
@@ -586,6 +586,11 @@ abstract class BaseReference extends BaseObject  implements Persistent
 			$deleteQuery = ReferenceQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(ReferencePeer::isIgnoringRights() || ReferencePeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -636,6 +641,11 @@ abstract class BaseReference extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(ReferencePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(ReferencePeer::isIgnoringRights() || ReferencePeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -653,6 +663,11 @@ abstract class BaseReference extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(ReferencePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(ReferencePeer::isIgnoringRights() || ReferencePeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
