+++ lib/model/om/BaseRight.php
@@ -864,6 +864,11 @@ abstract class BaseRight extends BaseObject  implements Persistent
 			$deleteQuery = RightQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(RightPeer::isIgnoringRights() || RightPeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -914,6 +919,11 @@ abstract class BaseRight extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(RightPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(RightPeer::isIgnoringRights() || RightPeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -931,6 +941,11 @@ abstract class BaseRight extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(RightPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(RightPeer::isIgnoringRights() || RightPeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
