+++ lib/model/om/BaseRole.php
@@ -496,6 +496,11 @@ abstract class BaseRole extends BaseObject  implements Persistent
 			$deleteQuery = RoleQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(RolePeer::isIgnoringRights() || RolePeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -546,6 +551,11 @@ abstract class BaseRole extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(RolePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(RolePeer::isIgnoringRights() || RolePeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -563,6 +573,11 @@ abstract class BaseRole extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(RolePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(RolePeer::isIgnoringRights() || RolePeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
