+++ lib/model/om/BaseString.php
@@ -525,6 +525,11 @@ abstract class BaseString extends BaseObject  implements Persistent
 			$deleteQuery = StringQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(StringPeer::isIgnoringRights() || StringPeer::mayOperateOn(Session::getSession()->getUser(), $this, "delete"))) {
+				throw new NotPermittedException("delete.custom", array("role_key" => ""));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -575,6 +580,11 @@ abstract class BaseString extends BaseObject  implements Persistent
 				if (!$this->isColumnModified(StringPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(StringPeer::isIgnoringRights() || StringPeer::mayOperateOn(Session::getSession()->getUser(), $this, "insert"))) {
+					throw new NotPermittedException("insert.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
@@ -592,6 +602,11 @@ abstract class BaseString extends BaseObject  implements Persistent
 				if ($this->isModified() && !$this->isColumnModified(StringPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
 				}
+				// denyable behavior
+				if(!(StringPeer::isIgnoringRights() || StringPeer::mayOperateOn(Session::getSession()->getUser(), $this, "update"))) {
+					throw new NotPermittedException("update.custom", array("role_key" => ""));
+				}
+
 				// attributable behavior
 				
 				if(Session::getSession()->isAuthenticated()) {
